Removed embedded SSE and made it external. Updated x360ce option.
SmartSteamEmu was embedded in the Nucleus.Gaming dll, which make it difficult to update or use a newer version if wanted. SSE is now an external file that can easily be updated/replaced by user or app. Greatly reduced Nucleus.Gaming dll size as well.

Also updated the x360ce option so it copies the respective xinput dll along with the x360ce executable.